fix: speed up unit tests
